Use eslintrc.js from styleguide
* Fix resulting Code Climate issues
* Extract IssueBuffer into support file
