+++ .codeclimate.yml
@@ -7,5 +7,5 @@ ratings:
   paths:
     - "**.js"
 exclude_paths:
-- "test"
-- "node_modules"
+- "node_modules/"
+- "test/fixtures/"
