+++ bin/fixme
@@ -5,6 +5,9 @@ var FixMe = require('../lib/fix-me');
 var config;
 
 fs.readFile('/config.json', function(err, data) {
-  if (!err) config = JSON.parse(data);
+  if (!err) {
+    config = JSON.parse(data);
+  }
+
   new FixMe().run(config)
 });
