+++ lib/fix-me.js
@@ -32,8 +32,8 @@ FixMe.prototype.run = function(engineConfig) {
     strings = DEFAULT_STRINGS;
   }
 
-  this.find(paths, strings)
-}
+  this.find(paths, strings);
+};
 
 FixMe.prototype.find = function(paths, strings, callback) {
   var pattern = `(${strings.join('|')})`;
@@ -61,7 +61,9 @@ FixMe.prototype.find = function(paths, strings, callback) {
     this.output.write(JSON.stringify(issue) + '\0');
   });
 
-  if (callback) grep.stdout.on('close', _ => callback());
-}
+  if (callback) {
+    grep.stdout.on('close', _ => callback());
+  }
+};
 
 module.exports = FixMe;
